Two resistors version
Because in the interdiction zone, the TL084 ount is not zero but 1.5v, I have
added a zenner diode at everey TL084's out. I have also added pull-down
resistors to get a more stable sistuation. But, in order to avoid to use a
bigger PCB, I have bive up to a resistor in the locker key. It reduce the
security level, at the end. But it remain hugh enough.